FIX: refactoring
