+++ src/pages/Detail/components/Filters/Line.js
@@ -1,9 +1,9 @@
 import React, { useContext, useState } from 'react';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 import { ReviewContext } from '../../Context';
-import High from './High';
-import Low from './Low';
+import High from './ByHigh';
+import Low from './ByLow';
 import ByDate from './ByDate';
 
 const Wrapper = styled.div`
@@ -18,7 +18,6 @@ function Line() {
   const { setFilter } = useContext(ReviewContext);
   const [now, setNow] = useState('high');
   const filterReviewComponent = e => {
-    // eslint-disable-next-line default-case
     switch (e.target.className) {
       case 'high': {
         setFilter(<High />);
@@ -35,14 +34,15 @@ function Line() {
         setNow('date');
         break;
       }
+      default: {
+        break;
+      }
     }
   };
 
   return (
-    // eslint-disable-next-line react/jsx-no-useless-fragment
     <Wrapper>
       {(function () {
-        // eslint-disable-next-line default-case
         switch (now) {
           case 'high': {
             return (
@@ -121,6 +121,33 @@ function Line() {
                   className="date"
                 >
                   √ 작성일순
+                </button>
+              </div>
+            );
+          }
+          default: {
+            return (
+              <div>
+                <button
+                  style={{ color: 'black' }}
+                  onClick={e => filterReviewComponent(e)}
+                  className="high"
+                >
+                  √ 높은평점순
+                </button>
+                <button
+                  style={{ color: 'silver' }}
+                  onClick={e => filterReviewComponent(e)}
+                  className="low"
+                >
+                  낮은평점순
+                </button>
+                <button
+                  style={{ color: 'silver' }}
+                  onClick={e => filterReviewComponent(e)}
+                  className="date"
+                >
+                  작성일순
                 </button>
               </div>
             );
