+++ src/pages/Detail/components/Modals/BenefitModal.js
@@ -53,7 +53,6 @@ function BenefitModal() {
     useContext(ModalContext);
 
   return (
-    // eslint-disable-next-line react/jsx-no-useless-fragment
     <div>
       {benefitModalOpen && (
         <Modal>
